Refactor: split monotonic.py into 3 files + shared helpers
Match bambi's one-class-per-file convention. The three Monotonic*Term
classes already covered three structurally distinct formulae representations
(standalone Term, interaction Term, GroupSpecificTerm) -- they're not
redundant -- but they had been crammed into one 453-line file while every
other term type lives in its own.

Layout now:
  bambi/terms/monotonic.py                 MonotonicTerm  (124 lines)
  bambi/terms/monotonic_interaction.py     MonotonicInteractionTerm  (124)
  bambi/terms/monotonic_group_specific.py  MonotonicGroupSpecificTerm  (155)
  bambi/terms/_monotonic_helpers.py        shared free functions  (89)

Shared helpers extracted into _monotonic_helpers.py:
* extract_component_info(call_component, idx=0) -> dict
    Pulls (transform, codes, D, K, levels, id, kind, idx) off a Monotonic
    stateful transform. Returns a dict (rather than a dataclass) because the
    predict path mutates it to cache re-encoded codes for new data.
* validate_prior_dict(value, allowed_keys, kind_label)
    Single implementation of the dict-shaped-prior validation that was
    duplicated across the three setters.
* simplex_names(mo_id, term_name) -> (name, dim)
    Centralizes the 'simplex_<id>' vs '{term_name}_simplex' naming used by
    all three classes.
* VALID_MONOTONIC_PRIOR_VALUES constant.

No intermediate base class between BaseTerm and the three subclasses, no
behavior change, no test change. All 33 monotonic tests still pass, all
210 other bambi tests still pass.
