+++ bambi/terms/__init__.py
@@ -1,11 +1,9 @@
 from bambi.terms.common import CommonTerm
 from bambi.terms.group_specific import GroupSpecificTerm
 from bambi.terms.hsgp import HSGPTerm
-from bambi.terms.monotonic import (
-    MonotonicTerm,
-    MonotonicInteractionTerm,
-    MonotonicGroupSpecificTerm,
-)
+from bambi.terms.monotonic import MonotonicTerm
+from bambi.terms.monotonic_group_specific import MonotonicGroupSpecificTerm
+from bambi.terms.monotonic_interaction import MonotonicInteractionTerm
 from bambi.terms.offset import OffsetTerm
 from bambi.terms.response import ResponseTerm
